# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Logueo",
        "description": "Endpoints para gestión de logueo",
        "contact": {
            "email": "email@gmail.com"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/retos": {
            "get": {
                "tags": [
                    "Retos"
                ],
                "summary": "Obtener todos los retos",
                "description": "Obtener todos los centros.",
                "operationId": "3e455cca76ddb74214b0c9525defa05f",
                "responses": {
                    "200": {
                        "description": "Lista de retos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "retos": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Reto"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Retos"
                ],
                "summary": "Crear un nuevo reto",
                "description": "Crear un nuevo centro.",
                "operationId": "fe7f9f9c7cd5c156b40164bd6c8490bb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Reto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Reto creado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Reto creado correctamente"
                                        },
                                        "reto": {
                                            "$ref": "#/components/schemas/Reto"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error en la validación de datos"
                    }
                }
            }
        },
        "/api/retos/{reto}": {
            "get": {
                "tags": [
                    "Retos"
                ],
                "summary": "Obtener un reto por su ID",
                "description": "Obtener un centro por su ID.",
                "operationId": "771e4d85276179007813223400266dc9",
                "parameters": [
                    {
                        "name": "reto",
                        "in": "path",
                        "description": "ID del reto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reto encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "reto": {
                                            "$ref": "#/components/schemas/Reto"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Reto no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Retos"
                ],
                "summary": "Actualizar un reto existente",
                "description": "Actualizar un reto existente.",
                "operationId": "97055e6aa10bcf0e3bc529cd02c6c30b",
                "parameters": [
                    {
                        "name": "reto",
                        "in": "path",
                        "description": "ID del reto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Reto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Reto actualizado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Reto actualizado correctamente"
                                        },
                                        "reto": {
                                            "$ref": "#/components/schemas/Reto"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error en la validación de datos"
                    },
                    "404": {
                        "description": "Reto no encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Retos"
                ],
                "summary": "Eliminar un reto",
                "description": "Eliminar un reto.",
                "operationId": "ab520caaaa757cef282b43855cc86adc",
                "parameters": [
                    {
                        "name": "reto",
                        "in": "path",
                        "description": "ID del reto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reto eliminado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Reto eliminado correctamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Reto no encontrado"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Reto": {
                "title": "Reto",
                "required": [
                    "titulo",
                    "texto",
                    "estudio_id"
                ],
                "properties": {
                    "id": {
                        "description": "ID único del reto",
                        "type": "integer",
                        "example": 1
                    },
                    "titulo": {
                        "description": "Titulo del reto",
                        "type": "string",
                        "example": "Primer gol del torneo"
                    },
                    "texto": {
                        "description": "Contenido del reto",
                        "type": "string",
                        "example": "Empezando todos los partidos a la vez el equipo que marque el primer gol gana el reto"
                    },
                    "estudio_id": {
                        "description": "ID relacionado con el estudio asociado al reto",
                        "type": "string",
                        "example": "3"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}